refactor: adjust utils exports
