+++ src/lib/utils/index.ts
@@ -1,8 +1,9 @@
+export * from './getCurrentTime';
+export * from './isNonNullable';
+export * from './isNull';
+export * from './isNumber';
+export * from './isString';
 export * from './isUndefined';
 export * from './noop';
-export * from './isString';
-export * from './isNumber';
-export * from './stableHash';
 export * from './objectKeys';
-export * from './getCurrentTime';
-export * from './waitFor';
+export * from './stableHash';
